Prefer variable syntax over function syntax over.
I don't know about you but I used to seeing that function syntax makes sense but after seeing other using the variable syntax and trying it, it actually made my code base less temtating and easier to read and reason with. I could find the right word to add to the this section, but a strong argument might be that Lua is based on the idea of tables, and often you see us defining a local table, setting default field values and exporting tables. It makes sense to used table syntax in lua for that reason.